Move textbox related component into its own folder
